+++ changelog.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v9.2.3...HEAD)
+### Changed
+* [STREAM-212](https://inindca.atlassian.net/browse/STREAM-212) - Bump streaming-client to pick up fix
+
 # [v9.2.3](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v9.2.2...v9.2.3)
 ### Changed
 * [NO_JIRA] - Bump client-logger to pick up fixes